refactoring code into functions
