+++ .gitignore
@@ -70,3 +70,6 @@ terraform.rc
 # Cache directories
 __pycache__/
 **/__pycache__/
+
+# Pycharm
+.idea/
