+++ code/python/flask_app/app/__init__.py
@@ -45,7 +45,6 @@ def create_app():
 
     return app
 
-
 def populate_initial_attribute_types():
     """
     Populate the database with initial data from archetypes.json.
