+++ code/python/flask_app/app/functions/openai.py
