+++ code/python/flask_app/app/functions/utils.py
