+++ code/python/flask_app/app/models.py
@@ -1,5 +1,4 @@
 from flask_sqlalchemy import SQLAlchemy
-from datetime import datetime
 
 # Define the DB object
 db = SQLAlchemy()
