--- code/python/flask_app/personas_payload.json
@@ -1,121 +1,0 @@
-{
-  "personas": [
-    {
-      "attributes": [
-        {
-          "id": "1",
-          "name": "economic",
-          "value": 0.2
-        },
-        {
-          "id": "2",
-          "name": "freedom",
-          "value": 0.85
-        },
-        {
-          "id": "3",
-          "name": "tone",
-          "value": 0.8
-        },
-        {
-          "id": "4",
-          "name": "cultural",
-          "value": 1
-        },
-        {
-          "id": "5",
-          "name": "conflict",
-          "value": 0.3
-        },
-        {
-          "id": "6",
-          "name": "optimism",
-          "value": 1
-        }
-      ],
-      "dob": "1990-06-15",
-      "id": "1",
-      "location": "New York",
-      "name": "Barry",
-      "profile_picture_s3_bucket_address": "https://example.com/images/barry.jpg"
-    },
-    {
-      "attributes": [
-        {
-          "id": "7",
-          "name": "economic",
-          "value": 0.5
-        },
-        {
-          "id": "8",
-          "name": "freedom",
-          "value": 0.5
-        },
-        {
-          "id": "9",
-          "name": "tone",
-          "value": 0.7
-        },
-        {
-          "id": "10",
-          "name": "cultural",
-          "value": 0.6
-        },
-        {
-          "id": "11",
-          "name": "conflict",
-          "value": 0.4
-        },
-        {
-          "id": "12",
-          "name": "optimism",
-          "value": 0.7
-        }
-      ],
-      "dob": "1985-04-20",
-      "id": "2",
-      "location": "Los Angeles",
-      "name": "Susan",
-      "profile_picture_s3_bucket_address": "https://example.com/images/susan.jpg"
-    },
-    {
-      "attributes": [
-        {
-          "id": "13",
-          "name": "economic",
-          "value": 0.3
-        },
-        {
-          "id": "14",
-          "name": "freedom",
-          "value": 0.9
-        },
-        {
-          "id": "15",
-          "name": "tone",
-          "value": 0.9
-        },
-        {
-          "id": "16",
-          "name": "cultural",
-          "value": 0.9
-        },
-        {
-          "id": "17",
-          "name": "conflict",
-          "value": 0.5
-        },
-        {
-          "id": "18",
-          "name": "optimism",
-          "value": 1
-        }
-      ],
-      "dob": "1995-08-10",
-      "id": "3",
-      "location": "Chicago",
-      "name": "Alex",
-      "profile_picture_s3_bucket_address": "https://example.com/images/alex.jpg"
-    }
-  ]
-}
