+++ code/python/flask_app/run_app.py
@@ -1,4 +1,3 @@
 from app import create_app
 
-# Create the application instance
 app = create_app()
